refactor: added icons with outlined
