--- packages/icons/src/ChevronUpIcon.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { IIconProps } from './types';
-
-export default (props: IIconProps): JSX.Element => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    className={props.className}
-    focusable={false}
-    aria-hidden="true"
-    viewBox="0 0 24 24"
-    fill="currentColor"
-  >
-    <path
-      fillRule="evenodd"
-      d="M14.707 12.707a1 1 0 01-1.414 0L10 9.414l-3.293 3.293a1 1 0 01-1.414-1.414l4-4a1 1 0 011.414 0l4 4a1 1 0 010 1.414z"
-      clipRule="evenodd"
-    />
-  </svg>
-);
