+++ packages/icons/src/ExpandLessIcon.tsx
@@ -10,6 +10,6 @@ export default (props: IIconProps): JSX.Element => (
     viewBox="0 0 24 24"
     fill="currentColor"
   >
-    <path d="M8 5a1 1 0 100 2h5.586l-1.293 1.293a1 1 0 001.414 1.414l3-3a1 1 0 000-1.414l-3-3a1 1 0 10-1.414 1.414L13.586 5H8zM12 15a1 1 0 100-2H6.414l1.293-1.293a1 1 0 10-1.414-1.414l-3 3a1 1 0 000 1.414l3 3a1 1 0 001.414-1.414L6.414 15H12z" />
+    <path d="m12 8-6 6 1.41 1.41L12 10.83l4.59 4.58L18 14z" />
   </svg>
 );
