+++ packages/icons/src/ExpandMoreIcon.tsx
@@ -10,10 +10,6 @@ export default (props: IIconProps): JSX.Element => (
     viewBox="0 0 24 24"
     fill="currentColor"
   >
-    <path
-      fillRule="evenodd"
-      d="M10 18a8 8 0 100-16 8 8 0 000 16zm1-11a1 1 0 10-2 0v2H7a1 1 0 100 2h2v2a1 1 0 102 0v-2h2a1 1 0 100-2h-2V7z"
-      clipRule="evenodd"
-    />
+    <path d="M16.59 8.59 12 13.17 7.41 8.59 6 10l6 6 6-6z" />
   </svg>
 );
