--- packages/icons/src/MinusCircle.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import { IIconProps } from './types';
-
-export default (props: IIconProps): JSX.Element => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    className={props.className}
-    focusable={false}
-    aria-hidden="true"
-    viewBox="0 0 24 24"
-    fill="currentColor"
-  >
-    <path
-      fillRule="evenodd"
-      d="M10 18a8 8 0 100-16 8 8 0 000 16zM7 9a1 1 0 000 2h6a1 1 0 100-2H7z"
-      clipRule="evenodd"
-    />
-  </svg>
-);
