--- packages/icons/src/MoonIcon.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { IIconProps } from './types';
-
-export default (props: IIconProps): JSX.Element => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    className={props.className}
-    focusable={false}
-    aria-hidden="true"
-    viewBox="0 0 24 24"
-    fill="currentColor"
-  >
-    <path d="M17.293 13.293A8 8 0 016.707 2.707a8.001 8.001 0 1010.586 10.586z" />
-  </svg>
-);
