+++ packages/icons/src/index.ts
@@ -1,16 +1,16 @@
 export { default as CancelIcon } from './CancelIcon';
 export { default as CheckCircleIcon } from './CheckCircleIcon';
-export { default as ChevronDownIcon } from './ChevronDownIcon';
-export { default as ChevronUpIcon } from './ChevronUpIcon';
+export { default as ExpandMoreIcon } from './ExpandMoreIcon';
+export { default as ExpandLessIcon } from './ExpandLessIcon';
 export { default as DangerousIcon } from './DangerousIcon';
-export { default as ExternalLinkIcon } from './ExternalLinkIcon';
+export { default as LaunchIcon } from './LaunchIcon';
 export { default as EmailIcon } from './EmailIcon';
 export { default as EyeIcon } from './EyeIcon';
 export { default as GearIcon } from './GearIcon';
 export { default as InfoIcon } from './InfoIcon';
-export { default as MinusCircle } from './MinusCircle';
-export { default as MoonIcon } from './MoonIcon';
-export { default as PlusCircleIcon } from './PlusCircleIcon';
+export { default as RemoveCircleIcon } from './RemoveCircleIcon';
+export { default as DarkModeIcon } from './DarkModeIcon';
+export { default as AddCircleIcon } from './AddCircleIcon';
 export { default as RotatingCircleIcon } from './RotatingCircleIcon';
-export { default as SunIcon } from './SunIcon';
+export { default as LightModeIcon } from './LightModeIcon';
 export { default as WarningIcon } from './WarningIcon';
